ManagerApp/requirements.txt

Flask
requests
python-dotenv
pymongo
redis

Faker
schedule

